refactor : notice 페이지 내 item 컴포넌트 분리 및 적용
